+++ visualization/app/codeCharta/ui/legendPanel/selectors/legendHeightMetric.selector.ts
@@ -1,10 +1,10 @@
 import { createSelector } from "../../../state/angular-redux/createSelector"
 import { heightMetricSelector } from "../../../state/store/dynamicSettings/heightMetric/heightMetric.selector"
 import { CcState } from "../../../state/store/store"
-import { metricDescriptions } from "../../../util/metric/metricDescriptions"
-import { LegendMetric } from "./legendMetric"
+import { getLegendMetric, LegendMetric } from "./legendMetric"
+import { attributeDescriptorsSelector } from "../../../state/store/fileSettings/attributeDescriptors/attributesDescriptors.selector"
 
-export const legendHeightMetricSelector: (state: CcState) => LegendMetric = createSelector([heightMetricSelector], heightMetric => ({
-	metricName: heightMetric,
-	description: metricDescriptions.get(heightMetric)
-}))
+export const legendHeightMetricSelector: (state: CcState) => LegendMetric = createSelector(
+	[heightMetricSelector, attributeDescriptorsSelector],
+	(heightMetric, attributeDescriptors) => getLegendMetric(heightMetric, attributeDescriptors)
+)